On branch Jaakko  Changes to be committed: 	modified:   .gitignore
Default Makefile
	new file:   Makefile_gnu

.txt output commented out
	modified:   bin/combine2.py
	modified:   bin/sound_in

Old version of data prosessing
	modified:   script/dataprosessointi.py

	modified:   script/submit_postpros.bash

Submit multiple runs
	modified:   script/testiajot.bash

Data prosessing methods
        new file:   script/ModDataPros.py

Data prosessing: LWP & RWP
	new file:   script/thermo4paths.py

Data prosessing caller function for thermo level 5
	new file:   script/thermo5datapros.py

Some debugging outputs
	modified:   src/src_LES/grid.f90
	modified:   src/src_LES/init.f90
	modified:   src/src_LES/ncio.f90
	modified:   src/src_LES/stat.f90
	modified:   src/src_LES/step.f90
	modified:   src/src_salsa/mo_salsa.f90
	modified:   src/src_salsa/mo_salsa_cloud.f90

+++ src/src_salsa/mo_salsa.f90
@@ -133,7 +133,7 @@ CONTAINS
     INTEGER, INTENT(in) :: level                         ! thermodynamical level
 
     INTEGER :: zpbl(kbdim)            ! Planetary boundary layer top level
-    !REAL(dp) :: zrh(kbdim,klev)           ! Relative humidity
+    !REAL :: zrh(kbdim,klev)           ! Relative humidity
 
 
     !-- Output variables -----------------------------------------------------------------
